refactor: restruct modules & add more methods  - move logic out of dataframe and into plan  - move all nested modules into their own top level file  - add logic to implement sort & orderby
