+++ Cargo.toml
@@ -18,12 +18,12 @@ include = [
 ]
 
 [dependencies]
-tonic = "0.10.0"
+tonic = { version ="0.11.0"}
 tokio = { version = "1.0", features = ["macros", "rt-multi-thread"] }
-tokio-rustls = "0.24.1"
+tokio-rustls = "0.25.0"
 
-arrow = { version = "46.0.0", features = ["prettyprint"] }
-arrow-ipc = "46.0.0"
+arrow = { version = "50.0.0", features = ["prettyprint"] }
+arrow-ipc = "50.0.0"
 
 prost = "0.12.0"
 prost-types = "0.12.0"
@@ -33,7 +33,7 @@ url = "2.2"
 
 
 [build-dependencies]
-tonic-build = "0.10.0"
+tonic-build = "0.11.0"
 
 [lib]
 doctest = false
