--- src/execution/mod.rs
@@ -1,3 +1,0 @@
-//! Create a Spark Session context from a Spark Connect remote source
-
-pub mod context;
