--- src/plan/mod.rs
@@ -1,6 +1,0 @@
-//! Logical Plan represents the gRPC Spark Plan used
-//! to create a DataFrame
-
-pub mod logical_plan;
-
-pub use logical_plan::LogicalPlanBuilder;
